chore: reorganize repo layout for clarity
- Move 12 smoke-test-*.mjs files from repo root into smoke/
- Update .github/workflows/deploy.yml paths-ignore (smoke/**)
- Rename src/chrome/ → src/layout/ (folder held a single Appbar.jsx)
- Update the one Appbar import in widgets/presentation/Presentation.jsx

Lint pre/post-reorganization shows identical output (2 errors, 3 warnings,
all pre-existing in useRuntime.js and LinkedListViz.jsx). BST smoke test
passes against dev server.
