+++ smoke/smoke-test-bst.mjs
