+++ smoke/smoke-test-embed.mjs
