+++ smoke/smoke-test-globalang.mjs
