+++ smoke/smoke-test-pages.mjs
