+++ smoke/smoke-test-presentation.mjs
