+++ smoke/smoke-test-stack.mjs
