+++ smoke/smoke-test-viz-nodes.mjs
