+++ smoke/smoke-test-viz.mjs
